fix(AENS): Change tld for Lima from `aet` to `chain` (#714)
* fix(AENS): Change tld for Lima from `aet` to `chain`
Point docker to the rc-5 node and 4.0.0 compiler
Move `classify` method from AENS to transaction builder helpers

* fix(AENS): Fix claim command with new tld
