+++ .env
@@ -1,2 +1,2 @@
-TAG=v5.0.0-rc.3
-COMPILER_TAG=v4.0.0-rc5
+TAG=v5.0.0-rc.5
+COMPILER_TAG=v4.0.0
